+++ crates/common/constants.rs
@@ -53,7 +53,7 @@ pub static DEPOSIT_TOPIC: LazyLock<H256> = LazyLock::new(|| {
 // === EIP-4844 constants ===
 
 /// Gas consumption of a single data blob (== blob byte size).
-pub const GAS_PER_BLOB: u64 = 1 << 17;
+pub const GAS_PER_BLOB: u32 = 1 << 17;
 
 // Minimum base fee per blob
 pub const MIN_BASE_FEE_PER_BLOB_GAS: u64 = 1;
