+++ crates/common/types/genesis.rs
@@ -95,8 +95,8 @@ impl TryFrom<&Path> for Genesis {
 )]
 #[serde(rename_all = "camelCase")]
 pub struct ForkBlobSchedule {
-    pub target: u64,
-    pub max: u64,
+    pub target: u32,
+    pub max: u32,
     pub base_fee_update_fraction: u64,
 }
